pyproject.toml
[project]
name = "qms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-decouple>=3.8",
    "ulid-py>=1.1.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.2",
]
